# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateReservationRequest:
    properties:
      expiresAt:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.ReservationRequestItemDTO'
        minItems: 1
        type: array
      orderId:
        type: string
    required:
    - expiresAt
    - items
    - orderId
    type: object
...
  dto.ReservationData:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ReservationItemData'
        type: array
      orderId:
        type: string
    type: object
  dto.ReservationItemData:
    properties:
      expiresAt:
        type: string
      qty:
        type: integer
      reservationId:
        type: string
      status:
        type: string
      variantId:
        type: string
    type: object
  dto.ReservationRequestItemDTO:
    properties:
      id:
        type: string
      qty:
        minimum: 1
        type: integer
    required:
    - id
    - qty
    type: object
  dto.ReservationSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ReservationData'
      status:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: 'REST API for Product & Category Service. Management routes require
    an Auth0 RS256 bearer access token whose exact permission matches the route; denials
    use the {"error":{"code":"...","message":"...","details":{}}} envelope with 401
    UNAUTHENTICATED for a missing, malformed, invalid, expired, or wrong-audience
    credential and 403 FORBIDDEN for a valid token without the route''s exact permission.
    Public exception: GET /categories, GET /categories/{id}, GET /categories/dropdown,
    GET /products, and GET /products/{id} require no token and ignore any supplied
    Authorization header. Customer exception: POST /products/{id}/reviews, PATCH /reviews/{reviewId},
    DELETE /reviews/{reviewId}, and GET /users/me/reviews keep the existing customer
    HS256 token and ownership rules, and an Auth0 permission is not a substitute for
    customer identity.'
  title: Gin Product Service API
  version: "1.0"
paths:
...
  /inventory/reservations:
    post:
      consumes:
      - application/json
      description: Reserve sufficient stock for one order at the default fulfillment
        location. Idempotent by orderId; the order service supplies the hold expiry.
      parameters:
      - description: Checkout reservation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReservationSuccessResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReservationSuccessResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an atomic checkout inventory reservation
      tags:
      - inventory
...
securityDefinitions:
  BearerAuth:
    description: Auth0 RS256 access token as `Bearer <token>`. The token must carry
      the operation's exact permission and be signed by a key currently published
      by the configured Auth0 domain. Token issuer and audience must match the configured
      Auth0 API; an invalid, expired, or unintended token yields 401.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"